Moves module chip into its own folder
